SEG-8351 Force enqueue del mutations on slave
When an open checkpoint is resend by the master to the slave,
the slave deletes the current instance of same checkpoint and
creates new one. Since the set of deletes which were resent are
already applied to hash table when they were sent first time,
those deletes are ignored and hence they wont appear in the
checkpoint mutations queue. This patch forces all deletes to be
enqueued in toWrite queue.