SEG-9083 Add token verification for performTapOp()
TapConnMap::performTapOp() uses tapconnection name for
performing tap operations. But, if somebody removes a tap
connection and adds a new tap connection with same name,
queued callbacks like background fetches would perform tap
operations on the new connection, which is having same name
as that of old connection.

This fix adds a TapConnection sessionID argument for performTapOp().
Hence by comparing token for callbacks and the tapconnection to
which the operation is going to be applied, we can identify whether
it is a valid operation on current tapconnection or not.
The sessionID is persistent for a TapProducer irrespective of
disconnects happening for tapconnection.