SEG-9577 ep_total_cache_size does not decrement correctly
The previous fix on this caused ep-engine to crash when a lot of non
resident items were set with new data. This was because of the way item
lengths are stored for non-resident items in dummy blobs. In this fix,
we subtract the difference between "the real total size of the item when
in memory" and "the key+blob size" whenever we evict items from memory.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>