Changed evictSize logic to evict till the requied allocated memory point is reached.
With different possible allocators used during compilation, the stat
observed for memory accounting during eviction will behave different.
For eg., tcmalloc does not free memory immediately, causing the stat to
remain about the same even when eviction is performed. This fix changes
the logic of evictSize: We measure the memory mark we need to reach and
evict till we get there or return false if we exhaust the queue. Also,
eviction happens in quanta, the size of which can be configured using
the eviction_quantum_size parameter; evictSize performs iterations of
evicting quanta of keys before checking if the desired memory mark is
reached.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>