Minimise requeueing items that already exist in the current eviction queue.
During eviction building, to avoid wastage of the primary queue when we
swap, we only build a secondary list and keep it ready. The front-end
will switch to this when it exhausts the primary list. Also,
lru_rebuild_stime forces a swap at the end of a queue build even if the
primary queue contains items (this is to ensure that the items on the
queue don't turn stale). In all other cases, the secondary queue
contains items with timestamps not lying in the range of the primary
queue so that we minimise overlap. All swaps happen under a lock and the
front-end doesn't wait to acquire the lock (it fails the operation).

Signed-off-by: Abhishek Nanda <ananda@zynga.com>