Build Fix: jemalloc_stats and rss.hh were fighting with each other!
rss.hh needed a separate rss.cc, this is in anticipation that rss.cc may
become a bit bigger with optimizations. Otherwise inlining GetSelfRSS()
in rss.hh would have sufficed to make build succeed.

Eliminated StoredValue::getAllocatedMemory and replaced it with
GetSelfRSS() to make it explicit that we are using RSS everywhere.

Signed-off-by: Vinay Y S <vinay@vys.in>