Fix: Race in flushOneDelOrSet vs PersistenceCallback
In case a set and a delete operation came for the same stored-value one
after another, we were queuing up two flushEntry items in case of this
race. That is, if the first flushEntry got requeued, we would end up with two
flushEntries. Then when first flushEntry is processed and stored-value
is deleted, second flushEntry will be pointing to dangling stored-value.

This is fixed by checking if the item is still clean in redirty function
and requeuing it only if it's still clean.

Signed-off-by: Vinay Y S <vinay@vys.in>