Optimization: flushOneDelOrSet simplificaiton
FIXME: needs further clean. This is a temporary commit for one
functional run.

Callbacks objects were being used with persistence layer. This was
unnecessary as the code sequence was really sequential. Also, the logic
inside the function had inconsistent handling for duplicate dirty
entries.

With this change, we now only delete a flushEntry if we are truly done
processing that particular stored-value. If the stored-value became
dirty while we were in the kvstore operation, when we come back we can
detect this and requeue the same stored-value.

Signed-off-by: Vinay Y S <vinay@vys.in>