fix: move scrub_memory to beginFlush
This will make sure that scrub_memory is called only when there's some
item available to flush. This is not a very good fix. The
FlusherHelper::run and Flusher::doFlush still keep ping-pong'ing between
each other. Will need a way to make them wait when ep_queue is empty to
fix this more correctly.

Signed-off-by: Vinay Y S <vinay@vys.in>