Recfactored queued time computation to a more logical code location
The value 'queued' is used only when the type of operation is
TAP_MUTATION, at which point the value is converted from uint32_t to
time_t by comparing the difference between the real system time with the
engine's current time. This fix moves this code into a location where
the TAP_MUTATION case is handled.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>